// src/app/models/student.model.ts
 export class Student {
     id: number;
    studentName: string;
    category: string;
    domicile: boolean;
    birthCertificate: boolean;
    marksheets: boolean;
    policeClearance: boolean;
    passport: boolean;
    declaration: boolean;
    dob: Date;
    fatherName: string;
    motherName: string;
    lastClassScore: number;
}